// src/app/interfaces/task.interface.ts
import { TaskDifficult, TaskLevel, TaskStatus } from "../constants/constants.enum";

export interface ITask {
    name: string;
    description: string;
    startDate: Date | string;
    dueDate: Date | string;
    status: TaskStatus;
    level: TaskLevel;
    image?: string;
    userId: string;
    difficulty: TaskDifficult
    resources?: object | [],
    id?: string;
}